feat: implement `Clone` for `TranslationOptions`
Since the struct does not implement the trait, it's difficult to reuse
the same options across multiple requests. The another way might include
to provided a custom `copy` or a `new` method that takes a reference to
the original struct, and creates a new with the same data. But since the
struct (1) does not contain any reference, (2) does not contain fields
which is hard or impossible to clone such as I/O resources, and (3) is a
fairly small so that the cost of clone is negligible.